Clean up redundant code constructs
(cherry picked from commit 028a70a296293e062277a849051e2616482eaa7b)

+++ core/src/se/tube42/drum/logic/EncodingService.java
@@ -99,7 +99,7 @@ public final class EncodingService
             return false;
 
         for(int i = 0; i < len; i++)
-            if( BASE64_STR.indexOf( (int)str.charAt(i)) < 0)
+            if( BASE64_STR.indexOf(str.charAt(i)) < 0)
                 return false;
 
         return true;
